[ADDED] Validation with theirs validators

+++ HRLeaveManagment.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using HRLeaveManagment.Application.Contracts.Persistence;
+using HRLeaveManagment.Application.Exceptions;
 using MediatR;
 
 namespace HRLeaveManagment.Application.Features.LeaveType.Commands.DeleteLeaveType
@@ -13,7 +14,13 @@ namespace HRLeaveManagment.Application.Features.LeaveType.Commands.DeleteLeaveTy
         }
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
+            var validator = new DeleteLeaveTypeCommandValidator();
+            var validationResult = await validator.ValidateAsync(request);
+            if (validationResult.Errors.Any())
+                throw new BadRequestException($"{request} is Invalid.", validationResult);
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+            if (leaveTypeToDelete == null)
+                throw new NotFoundException(nameof(LeaveType), request.Id);
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
             return Unit.Value;
         }
